% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dresden_palette.R
\name{dresden_palette}
\alias{dresden_palette}
\title{A Dresden Files palette generator}
\usage{
dresden_palette(title, n, type = c("discrete", "continuous", "paired"))
}
\arguments{
\item{n:}{Number of colors desired. Default uses all colors.}

\item{title:}{Name of color palette (Dresden Files book title) desired.}

\item{type:}{Either "discrete", "continuous", or "paired".}
}
\description{
This function allows you to generate color palettes based on Dresden Files cover art.
}
\examples{
dresden_palette()
}
\keyword{color}
\keyword{palettes}
