% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dresden_palettes.R
\docType{data}
\name{dresden_palettes}
\alias{dresden_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 18.}
\usage{
dresden_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}
