% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Function to plot the color palettes}
\usage{
plot_palette(pal, title)
}
\arguments{
\item{n:}{Number of colors desired. Default uses all colors.}

\item{title:}{Name of color palette (Dresden Files book title) desired.}
}
\description{
Function to plot the color palettes
}
